refactor: use tiptapEditor component in modele form
