+++ gsl_core/templates/base.html
@@ -47,8 +47,8 @@
             {
                 "imports": {
                     "stimulus": "/static/vendor/stimulus.js",
-                    "modal": "/static/ui/components/modal.js",
-                    "tiptapEditor": "/static/ui/components/tiptapEditor.js",
+                    "modal": "/static/ui/controllers/modal.js",
+                    "tiptapEditor": "/static/ui/controllers/tiptapEditor.js",
                     "doubleDotation": "/static/js/controllers/doubleDotation.js"
                 }
             }
