+++ gsl_notification/templates/gsl_notification/modele_arrete/modelearrete_form_base.html
@@ -5,7 +5,7 @@
     <link rel="stylesheet" href="{% static "css/gsl.css" %}">
     <link rel="stylesheet" href="{% static "css/programmation_projet_list.css" %}">
     <link rel="stylesheet" href="{% static "css/notification.css" %}">
-    <link rel="stylesheet" href="{% static 'css/tiptap.css' %}">
+    <link rel="stylesheet" href="{% static 'ui/components/tiptap.css' %}">
 {% endblock extra_css %}
 
 {% block extra_js %}
