+++ gsl_notification/templates/gsl_notification/tab_simulation_projet/tab_notifications.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="{% static 'css/gsl.css' %}">
     <link rel="stylesheet" href="{% static 'css/projet_detail.css' %}">
     <link rel="stylesheet" href="{% static 'css/notification.css' %}">
-    <link rel="stylesheet" href="{% static 'css/tiptap.css' %}">
+    <link rel="stylesheet" href="{% static 'ui/components/tiptap.css' %}">
 {% endblock extra_css %}
 {% block tab_projet %}
     <h1 class="fr-mb-8w fr-mr-10w">
