+++ ui/static/ui/components/tiptap.css
